Hash problems 1 (#34)
This adds a few example problems that are tagged as "hash table" (all from Leetcode). We reviewed the first one (keyboard-row) in the online workshop, and so I added a discussion of common gotchas or trip-ups I saw when observing.

I highlighted these problems in particular because I believe they reflect real-world loops and conditions, though of course the surface-level motivation is contrived.